refactor(deploy): rename HasDeployConfig trait to HasDeployConfigTrait across all deploy commands
- add explicit HasDeployConfigTrait import statement to all deploy command classes
- change trait usage from HasDeployConfig to HasDeployConfigTrait in all command classes
- update DbBackupCommand, DbCreateCommand, DbRestoreCommand, DeployCommand, EnvPullCommand
- update LogsTailCommand, LogsViewCommand, ProvisionCommand, RollbackCommand
- update ScheduleRemoveCommand, ScheduleSetupCommand, ScheduleStatus
