+++ src/Framework/Deploy/Commands/DbCreateCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 
 /**
  * Create a new MySQL database and user on a remote server.
@@ -21,7 +22,7 @@ use Lightpack\Console\Command;
  */
 class DbCreateCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
