+++ src/Framework/Deploy/Commands/EnvPullCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 use Lightpack\Utils\Process;
 
 /**
@@ -17,7 +18,7 @@ use Lightpack\Utils\Process;
  */
 class EnvPullCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
