+++ src/Framework/Deploy/Commands/LogsTailCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 
 /**
  * Stream application logs in real-time from a remote server.
@@ -13,7 +14,7 @@ use Lightpack\Console\Command;
  */
 class LogsTailCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
