+++ src/Framework/Deploy/Commands/LogsViewCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 
 /**
  * View the last N lines of application logs on a remote server.
@@ -14,7 +15,7 @@ use Lightpack\Console\Command;
  */
 class LogsViewCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
