+++ src/Framework/Deploy/Commands/ServerQueueListCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 use Lightpack\Utils\Process;
 
 /**
@@ -13,7 +14,7 @@ use Lightpack\Utils\Process;
  */
 class ServerQueueListCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
