+++ src/Framework/Deploy/Commands/ServerQueueRestartCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 
 /**
  * Restart the supervised queue worker on a remote server.
@@ -13,7 +14,7 @@ use Lightpack\Console\Command;
  */
 class ServerQueueRestartCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
