+++ src/Framework/Deploy/Commands/ServerQueueStartCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 
 /**
  * Start the supervised queue worker on a remote server.
@@ -13,7 +14,7 @@ use Lightpack\Console\Command;
  */
 class ServerQueueStartCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
