+++ src/Framework/Deploy/Commands/SiteAddCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 
 /**
  * Add an Nginx virtual host for a domain.
@@ -12,7 +13,7 @@ use Lightpack\Console\Command;
  */
 class SiteAddCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
