+++ src/Framework/Deploy/Commands/SslCommand.php
@@ -3,6 +3,7 @@
 namespace Lightpack\Deploy\Commands;
 
 use Lightpack\Console\Command;
+use Lightpack\Deploy\HasDeployConfigTrait;
 
 /**
  * Obtain and install an SSL certificate via Certbot.
@@ -12,7 +13,7 @@ use Lightpack\Console\Command;
  */
 class SslCommand extends Command
 {
-    use HasDeployConfig;
+    use HasDeployConfigTrait;
 
     public function run()
     {
